Got rid of byte data type in favor of uint (and double), stars limit to 5

+++ eshopBackend.API/Controllers/CategoriesController.cs
@@ -21,7 +21,7 @@ public class CategoriesController : ControllerBase
 
 
     [HttpGet("list/{page}")]
-    public List<CategoryEntity>? GetCategories(byte page)
+    public List<CategoryEntity>? GetCategories(uint page)
     {
         List<CategoryEntity>? categories = _categoryRepository.CategoriesOverview(page);
         return categories;
